API Documentation

API documentation

Hermes functions

These functions publish and subscribe to Hermes MQTT-topics.

Main.SnipsHermesQnD.subscribe2Intents — Function.

subscribe2Intents(intents, callback; moreTopics = nothing)

Subscribe to one or a list of Intents and listen forever and run the callback if a matching intent is recieved.

Arguments:

intents: Abstract String or List of Abstract Strings to define intents to subscribe. The intents will be expanded to topics (i.e. "hermes/intent/SwitchOnLight")

callback: Function to be executed for a incoming message

moreTopics: keyword arg to provide additional topics to subscribe (complete names of of topics).

Details:

The callback function has the signature f(topic, intentMessage), where topic is a String and intentMessage a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not valid JSON. The callback function is not spawned, but executed in the current thread. As a result the function is not listening during execution of the callback.

source

Main.SnipsHermesQnD.publishStartSessionAction — Function.

publishStartSessionAction(text; siteId = CURRENT_SITE_ID,
     intentFilter = nothing, sendIntentNotRecognized = false,
     customData = nothing)

MQTT publish start session with init action

Arguments:

siteId: ID of the session to be terminated as String

text: text to be said via TTS

intentFilter: Optional Array of String - A list of intents names to restrict the NLU resolution on the answer of this query.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

customData: data to be sent to the service.

source

Main.SnipsHermesQnD.publishStartSessionNotification — Function.

publishStartSessionNotification(text; siteId = CURRENT_SITE_ID,
                                customData = nothing)

MQTT publish start session with init notification

Arguments:

siteId: ID of the session to be terminated as String

text: text to be said via TTS

customData: data to be sent to the service.

source

Main.SnipsHermesQnD.publishEndSession — Function.

publishEndSession(text; sessionId = CURRENT_SESSION_ID)

MQTT publish end session.

Arguments:

sessionId: ID of the session to be terminated as String If omiited, sessionId of the current will be inserted.

text: text to be said via TTS

source

Dialogue manager functions

In addition to functions to work with the dialogue manager advanced direct dialogues are provided that can be included in the control flow of the program.

Main.SnipsHermesQnD.publishContinueSession — Function.

publishContinueSession(text; sessionId = CURRENT_SESSION_ID,
     intentFilter = nothing,
     customData = nothing, sendIntentNotRecognized = false)

MQTT publish continue session.

Arguments:

sessionId: ID of the session to be terminated as String

text: text to be said via TTS

intentFilter: Optional Array of String - A list of intents names to restrict the NLU resolution on the answer of this query.

customData: Optional String - an update to the session's custom data.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

source

Main.SnipsHermesQnD.listenIntentsOneTime — Function.

listenIntentsOneTime(intents; moreTopics = nothing)

Subscribe to one or a list of Intents, but listen only until one matching intent is recognised.

Arguments

intents: Abstract String or List of Abtsract Strings to define intents to subscribe or nothing

moreTopics: keyword arg to provide additional topics to subscribe,

Value:

Return values are topic (as String) and payload (as Dict or as String if JSON parsing is not possible). If the topic is an intent, only the intent id is returned (i.e.: devname:intentname without the leading hermes/intent/)

source

Main.SnipsHermesQnD.configureIntent — Function.

configureIntent(intent, on)

Enable or disable an intent.

Arguments:

intent: one intent to be configured

on: boolean value; if true, the intent is enabled; if false it is disabled.

source

Main.SnipsHermesQnD.askYesOrNoOrUnknown — Function.

askYesOrNoOrUnknown(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :yes if "Yes" is answered or :no if "No" or :unknown otherwise.

Arguments:

question: String with the question to be uttered by Snips

source

Main.SnipsHermesQnD.askYesOrNo — Function.

askYesOrNo(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :true if "Yes" or "No" otherwise.

Arguments:

question: String with the question to uttered

source

Main.SnipsHermesQnD.publishSay — Function.

publishSay(text; sessionId = CURRENT_SESSION_ID,
                lang = nothing, id = nothing, wait = true)

Let the TTS say something.

Arguments:

text: text to be said vie TTS

lang: Optional language code to use when saying the text. If nothing is provided, en will be used

sessionId: optional ID of the session if there is one

id: optional A request identifier. If provided, it will be passed back in the response on hermes/tts/sayFinished.

wait: wait until the massege is spoken (i.i. wait for the MQTT-topic)

source

config.ini functions

Helper functions for read values from the file config.ini.

config.ini files follow the normal rules as for all Snips apps, with one extension:

no spaces around the =

if value of the parameter (right side) includes commas, the value can be interpreted as a comma-separated list of values. In this case, the reader-function will return an array of Strings with the values (which an be accessed by their index).

Main.SnipsHermesQnD.readConfig — Function.

readConfig(appDir)

Read the lines of the App's config file and return a Dict with config values.

Arguments:

appDir: Directory of the currently running app.

source

Main.SnipsHermesQnD.matchConfig — Function.

matchConfig(name::Symbol, val::String)

Return true if the parameter with name name of the config.ini has the value val or one element of the list as the value val.

Arguments:

name: name of the config parameter as Symbol

val: desired value

source

Main.SnipsHermesQnD.getConfig — Function.

getConfig(name)

Return the parameter value of the config.ini with name or nothing if the param does not exist. Return value is of type AbstractString, if it is a single value or of type AbstractArray{AbstractString} if the a list of values is read.

Arguments:

name: name of the config parameter as Symbol or String

source

Missing docstring.

Missing docstring for isinConfig. Check Documenter's build log for details.

Main.SnipsHermesQnD.getAllConfig — Function.

getAllConfig()

Return a Dict with the complete config.ini.

source

Slot access functions

Functions to read values from slots of recognised intents.

Main.SnipsHermesQnD.extractSlotValue — Function.

extractSlotValue(payload, slotName; multiple = false)

Return the value of a slot.

Nothing is returned, if

no slots in payload,

no slots with name slotName in payload,

no values in slot slotName.

If multiple == true, a list of all slot values will be returned. If false, only the 1st one as String.

source

Main.SnipsHermesQnD.isInSlot — Function.

isInSlot(payload, slotName, value)

Return true, if the value is present in the slot slotName of the JSON payload (i.e. one of the slot values must match). Return false if something is wrong (value not in payload or no slots slotName.)

source

Main.SnipsHermesQnD.isOnOffMatched — Function.

isOnOffMatched(payload, deviceName; siteId = CURRENT_SITE_ID)

Action to be combined with the ADoSnipsOnOFF intent. Depending on the payload the function returns:

:on if "on"

:off if "off"

:matched, if the device is matched but no on or off

:unmatched, if one of

wrong siteId

wrong device

Arguments:

payload: payload of intent

siteId: siteId of the device to be matched with the payload of intent

deviceName : name of device to be matched with the payload of intent

source

MQTT functions

Low-level API to MQTT messages (publish and subscribe). In the QuickAndDirty framework, these functions are calling Eclipse mosquitto_pub and mosquitto_sub. However this first (and preliminary) implementation is surpriningly robust and easy to maintain - it seems there is no need to change.

Main.SnipsHermesQnD.subscribeMQTT — Function.

subscribeMQTT(topics, callback; hostname = nothing, port = nothing)

Listen to one or more topics.

Arguments

topics: Abstract String or List of Abtsract Strings to define topics to subscribe

callback: Function to be executed for a incoming message.

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

Details:

The callback function has the signature f(topic, payload), where topic is a String and payload a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-formate) or a String, if the payload is not valid JSON.

The callback function is not spawned, but executed in the current thread. As a result the function is not listening during execution of the callback.

source

Main.SnipsHermesQnD.readOneMQTT — Function.

readOneMQTT(topics; hostname = nothing, port = nothing)

Listen to one or more topics until one message is retrieved and return topic as string and payload as Dict or as String if JSON parsing is not possible).

Arguments

topics: Abstract String or List of Abtsract Strings to define topics to subscribe

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

source

Main.SnipsHermesQnD.publishMQTT — Function.

publishMQTT(topic, payload, hostname = nothing, port = nothing)

Publishe a MQTT message.

Arguments

topics: String with the topic

payload: Dict() with message

hostname:

port: Hostname and port to use. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

source

Utility functions

Little helpers to provide functionality whichis commonly needed when develloping a skill.

Handle siteId from recognised intent

Main.SnipsHermesQnD.setSiteId — Function.

setSiteId(siteId)

Set the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

source

Main.SnipsHermesQnD.getSiteId — Function.

getSiteId()

Return the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

source

Other Utils

Main.SnipsHermesQnD.setLanguage — Function.

setLanguage(lang)

Set the default language for SnipsHermesQnD.

source

Main.SnipsHermesQnD.tryrun — Function.

tryrun(cmd; wait = true, errorMsg = TEXTS[:error_script])

Try to run an external command and returns true if successful or false if not.

source

Main.SnipsHermesQnD.tryReadTextfile — Function.

tryReadTextfile(fname, errMsg = TEXTS[:error_read])

Try to read a text file from file system and return the text as String or an String of length 0, if something went wrong.

source

Main.SnipsHermesQnD.tryParseJSONfile — Function.

tryParseJSONfile(fname)

Parse a JSON file and return a hierarchy of Dicts with the content.

keys are changed to Symbol

if error, nothing is returned

source

Main.SnipsHermesQnD.tryParseJSON — Function.

tryParseJSON(text)

parses a JSON and returns a hierarchy of Dicts{Symbol, Any} and Arrays with the content or a string (text), if text is not a valid JSON string is returned.

source

Main.SnipsHermesQnD.tryMkJSON — Function.

tryMkJSON(payload)

Create a JSON representation of the input (nested Dict or Array) and return an empty string if not possible.

source

Main.SnipsHermesQnD.setGPIO — Function.

setGPIO(gpio, onoff::Symbol)

Switch a GPIO on or off.

Arguments:

gpio: ID of GPIO (not pinID)

onoff: one of :on or :off

source

Index

Main.SnipsHermesQnD.askYesOrNo

Main.SnipsHermesQnD.askYesOrNoOrUnknown

Main.SnipsHermesQnD.configureIntent

Main.SnipsHermesQnD.extractSlotValue

Main.SnipsHermesQnD.getAllConfig

Main.SnipsHermesQnD.getConfig

Main.SnipsHermesQnD.getSiteId

Main.SnipsHermesQnD.isInSlot

Main.SnipsHermesQnD.isOnOffMatched

Main.SnipsHermesQnD.listenIntentsOneTime

Main.SnipsHermesQnD.matchConfig

Main.SnipsHermesQnD.publishContinueSession

Main.SnipsHermesQnD.publishEndSession

Main.SnipsHermesQnD.publishMQTT

Main.SnipsHermesQnD.publishSay

Main.SnipsHermesQnD.publishStartSessionAction

Main.SnipsHermesQnD.publishStartSessionNotification

Main.SnipsHermesQnD.readConfig

Main.SnipsHermesQnD.readOneMQTT

Main.SnipsHermesQnD.setGPIO

Main.SnipsHermesQnD.setLanguage

Main.SnipsHermesQnD.setSiteId

Main.SnipsHermesQnD.subscribe2Intents

Main.SnipsHermesQnD.subscribeMQTT

Main.SnipsHermesQnD.tryMkJSON

Main.SnipsHermesQnD.tryParseJSON

Main.SnipsHermesQnD.tryParseJSONfile

Main.SnipsHermesQnD.tryReadTextfile

Main.SnipsHermesQnD.tryrun