Dialogues

Dialogue manager functions

In addition to functions to work with the dialogue manager advanced direct dialogues are provided that can be included in the control flow of the program.

Main.SnipsHermesQnD.publishContinueSession — Function.

publishContinueSession(text; sessionId = CURRENT_SESSION_ID,
     intentFilter = nothing,
     customData = nothing, sendIntentNotRecognized = false)

MQTT publish continue session.

Arguments:

sessionId: ID of the session to be terminated as String

text: text to be said via TTS

intentFilter: Optional Array of String - A list of intents names to restrict the NLU resolution on the answer of this query.

customData: Optional String - an update to the session's custom data.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

source

Main.SnipsHermesQnD.listenIntentsOneTime — Function.

listenIntentsOneTime(intents; moreTopics = nothing)

Subscribe to one or a list of Intents, but listen only until one matching intent is recognised.

Arguments

intents: Abstract String or List of Abtsract Strings to define intents to subscribe or nothing

moreTopics: keyword arg to provide additional topics to subscribe,

Value:

Return values are topic (as String) and payload (as Dict or as String if JSON parsing is not possible). If the topic is an intent, only the intent id is returned (i.e.: devname:intentname without the leading hermes/intent/)

source

Main.SnipsHermesQnD.configureIntent — Function.

configureIntent(intent, on)

Enable or disable an intent.

Arguments:

intent: one intent to be configured

on: boolean value; if true, the intent is enabled; if false it is disabled.

source

Main.SnipsHermesQnD.askYesOrNoOrUnknown — Function.

askYesOrNoOrUnknown(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :yes if "Yes" is answered or :no if "No" or :unknown otherwise.

Arguments:

question: String with the question to be uttered by Snips

source

Main.SnipsHermesQnD.askYesOrNo — Function.

askYesOrNo(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :true if "Yes" or "No" otherwise.

Arguments:

question: String with the question to uttered

source

Main.SnipsHermesQnD.publishSay — Function.

publishSay(text; sessionId = CURRENT_SESSION_ID,
                lang = nothing, id = nothing, wait = true)

Let the TTS say something.

Arguments:

text: text to be said vie TTS

lang: Optional language code to use when saying the text. If nothing is provided, en will be used

sessionId: optional ID of the session if there is one

id: optional A request identifier. If provided, it will be passed back in the response on hermes/tts/sayFinished.

wait: wait until the massege is spoken (i.i. wait for the MQTT-topic)

source