Hermes

Hermes functions

These functions publish and subscribe to Hermes MQTT-topics.

Main.SnipsHermesQnD.subscribe2Intents — Function.

subscribe2Intents(intents, callback; moreTopics = nothing)

Subscribe to one or a list of Intents and listen forever and run the callback if a matching intent is recieved.

Arguments:

intents: Abstract String or List of Abstract Strings to define intents to subscribe. The intents will be expanded to topics (i.e. "hermes/intent/SwitchOnLight")

callback: Function to be executed for a incoming message

moreTopics: keyword arg to provide additional topics to subscribe (complete names of of topics).

Details:

The callback function has the signature f(topic, intentMessage), where topic is a String and intentMessage a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not valid JSON. The callback function is not spawned, but executed in the current thread. As a result the function is not listening during execution of the callback.

source

Main.SnipsHermesQnD.publishStartSessionAction — Function.

publishStartSessionAction(text; siteId = CURRENT_SITE_ID,
     intentFilter = nothing, sendIntentNotRecognized = false,
     customData = nothing)

MQTT publish start session with init action

Arguments:

siteId: ID of the session to be terminated as String

text: text to be said via TTS

intentFilter: Optional Array of String - A list of intents names to restrict the NLU resolution on the answer of this query.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

customData: data to be sent to the service.

source

Main.SnipsHermesQnD.publishStartSessionNotification — Function.

publishStartSessionNotification(text; siteId = CURRENT_SITE_ID,
                                customData = nothing)

MQTT publish start session with init notification

Arguments:

siteId: ID of the session to be terminated as String

text: text to be said via TTS

customData: data to be sent to the service.

source

Main.SnipsHermesQnD.publishEndSession — Function.

publishEndSession(text; sessionId = CURRENT_SESSION_ID)

MQTT publish end session.

Arguments:

sessionId: ID of the session to be terminated as String If omiited, sessionId of the current will be inserted.

text: text to be said via TTS

source