config.ini

config.ini functions

Helper functions for read values from the file config.ini.

config.ini files follow the normal rules as for all Snips apps, with one extension:

no spaces around the =

if value of the parameter (right side) includes commas, the value can be interpreted as a comma-separated list of values. In this case, the reader-function will return an array of Strings with the values (which an be accessed by their index).

Main.SnipsHermesQnD.readConfig — Function.

readConfig(appDir)

Read the lines of the App's config file and return a Dict with config values.

Arguments:

appDir: Directory of the currently running app.

source

Main.SnipsHermesQnD.matchConfig — Function.

matchConfig(name::Symbol, val::String)

Return true if the parameter with name name of the config.ini has the value val or one element of the list as the value val.

Arguments:

name: name of the config parameter as Symbol

val: desired value

source

Main.SnipsHermesQnD.getConfig — Function.

getConfig(name)

Return the parameter value of the config.ini with name or nothing if the param does not exist. Return value is of type AbstractString, if it is a single value or of type AbstractArray{AbstractString} if the a list of values is read.

Arguments:

name: name of the config parameter as Symbol or String

source

Main.SnipsHermesQnD.isinConfig — Function.

isinConfig(name::Symbol)

Return true if a parameter with name exists.

Arguments:

* `name`: name of the config parameter as Symbol

source

Main.SnipsHermesQnD.getAllConfig — Function.

getAllConfig()

Return a Dict with the complete config.ini.

source