MQTT

MQTT functions

Low-level API to MQTT messages (publish and subscribe). In the QuickAndDirty framework, these functions are calling Eclipse mosquitto_pub and mosquitto_sub. However this first (and preliminary) implementation is surpriningly robust and easy to maintain - it seems there is no need to change.

Main.SnipsHermesQnD.subscribeMQTT — Function.

subscribeMQTT(topics, callback; hostname = nothing, port = nothing)

Listen to one or more topics.

Arguments

topics: Abstract String or List of Abtsract Strings to define topics to subscribe

callback: Function to be executed for a incoming message.

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

Details:

The callback function has the signature f(topic, payload), where topic is a String and payload a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-formate) or a String, if the payload is not valid JSON.

The callback function is not spawned, but executed in the current thread. As a result the function is not listening during execution of the callback.

source

Main.SnipsHermesQnD.readOneMQTT — Function.

readOneMQTT(topics; hostname = nothing, port = nothing)

Listen to one or more topics until one message is retrieved and return topic as string and payload as Dict or as String if JSON parsing is not possible).

Arguments

topics: Abstract String or List of Abtsract Strings to define topics to subscribe

hostname:

port: Hostname and port to listen. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

source

Main.SnipsHermesQnD.publishMQTT — Function.

publishMQTT(topic, payload, hostname = nothing, port = nothing)

Publishe a MQTT message.

Arguments

topics: String with the topic

payload: Dict() with message

hostname:

port: Hostname and port to use. If not specified mosquitto_sub will be called without hostname/port (use the default configuration of the system).

source