Slots

Slot access functions

Functions to read values from slots of recognised intents.

Main.SnipsHermesQnD.extractSlotValue — Function.

extractSlotValue(payload, slotName; multiple = false)

Return the value of a slot.

Nothing is returned, if

no slots in payload,

no slots with name slotName in payload,

no values in slot slotName.

If multiple == true, a list of all slot values will be returned. If false, only the 1st one as String.

source

Main.SnipsHermesQnD.isInSlot — Function.

isInSlot(payload, slotName, value)

Return true, if the value is present in the slot slotName of the JSON payload (i.e. one of the slot values must match). Return false if something is wrong (value not in payload or no slots slotName.)

source

Main.SnipsHermesQnD.isOnOffMatched — Function.

isOnOffMatched(payload, siteId, deviceName)

Action to be combined with the ADoSnipsOnOFF intent. Depending on the payload the function returns:

:on if "on"

:off

:matched, if the device is matched but no on or off

:unmatched, if one of

wrong siteId

wrong device

Arguments:

payload: payload of intent

siteId: siteId of the device to be matched with the payload of intent

deviceName : name of device to be matched with the payload of intent

source