Utilities

Utility functions

Little helpers to provide functionality whichis commonly needed when develloping a skill.

Handle siteId from recognised intent

Main.SnipsHermesQnD.setSiteId — Function.

setSiteId(siteId)

Set the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

source

Main.SnipsHermesQnD.getSiteId — Function.

getSiteId()

Return the siteId in the Module SnipsHermesQnD (necessary to direct the say() output to the current room)

source

Other Utils

Main.SnipsHermesQnD.setLanguage — Function.

setLanguage(lang)

Set the default language for SnipsHermesQnD.

source

Main.SnipsHermesQnD.tryrun — Function.

tryrun(cmd; wait = true, errorMsg = TEXTS[:error_script])

Try to run an external command and returns true if successful or false if not.

source

Main.SnipsHermesQnD.tryReadTextfile — Function.

tryReadTextfile(fname, errMsg = TEXTS[:error_read])

Try to read a text file from file system and return the text as String or an String of length 0, if something went wrong.

source

Main.SnipsHermesQnD.tryParseJSONfile — Function.

tryParseJSONfile(fname)

Parse a JSON file and return a hierarchy of Dicts with the content.

keys are changed to Symbol

if error, nothing is returned

source

Main.SnipsHermesQnD.tryParseJSON — Function.

tryParseJSON(text)

parses a JSON and returns a hierarchy of Dicts{Symbol, Any} and Arrays with the content or a string (text), if text is not a valid JSON string is returned.

source

Main.SnipsHermesQnD.tryMkJSON — Function.

tryMkJSON(payload)

Create a JSON representation of the input (nested Dict or Array) and return an empty string if not possible.

source

Main.SnipsHermesQnD.setGPIO — Function.

setGPIO(gpio, onoff::Symbol)

Switch a GPIO on or off.

Arguments:

gpio: ID of GPIO (not pinID)

onoff: one of :on or :off

source